signtalk/requirements.txt
# Django Web框架依赖
Django==4.2.17
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
PyMySQL==1.1.0

# 手语识别模型依赖
absl-py==2.1.0
attrs==24.3.0
cffi==1.17.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
Jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==10.4.0
protobuf==3.20.3
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.10.1
six==1.17.0
sounddevice==0.5.1
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
zipp==3.20.2

# 其他常用依赖
Pillow==10.4.0
requests==2.31.0




